pyproject.toml
[tool.poetry]
name = "homepage"
version = "0.1.0"
description = ""
authors = ["Sina Siadat <siadat@gmail.com>"]

[tool.poetry.dependencies]
python = "^2.7"
Jinja2 = "^2.10"
markdown = "^3.1"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
